+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Updated Ruby to version 2.6.1 (#3452)
 - Changed custom BonApp cafe viewer icon to a cog instead of the ionicons logo (#3458)
 - Updated `react-native-vector-icons` to v6 and made some compatibility fixes (#3162)
+- Addressed color banding in SIS/Balances on Android (#3462)
 
 ### Fixed
 - Fixed an issue where Fastlane was reporting build failures despite having skipped the build (#3215)